Add Nvidia-Docker support
Added Dockerfile and .dockerignore to build an image with all
dependencies to compile and run grmonty in any machine with a GPU,
nvidia-docker installation and GPU drivers. Modified .gitignore to
include pgilinux*.tar.gz (necessary for the docker image). Added all
docker information in README.md. Also, renamed makefile to Makefile for
standardization.