Builds and run tests on a temp dir
Changed Makefile to accept an environment variable defining where to
build grmonty. Changed test/tester.py so that it creates a temp dir to
build grmonty and run tests. This dir is deleted after tester execution.
Now, tester.py exists gracefully when a SIGINT is sent to it. Updated
test/README.md and .gitignore for these new changes.