Convert references file to json and ajust settings
Previously, references were stored as a python script file that was
imported by test/tester.py. This was not a good approach since .py
files were not made to store data but code, so test/tester.py was
adapted to save and read from a .json file. Also some default settings
were adjusted.