Disable compilation during OpenACC to CUDA convertion
This commit starts a series of commits whose purpose is to convert the
OpenACC code to CUDA. This commit disables grmonty compilation and,
instead, when one tries to run make, it prints a nice message. This is
done because, for some commits, until convesion is mature, the
compilation will be broken.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>