Track photons in batches
This commit refactores the core loop to send and track photons in
batches (sending the next batch async at the same time that the previous
batch sent is beeing tracked).
This is the first commit in the series that aims to truncate the
simulation of photons that take too long, to be simulated a posteriori.

Note: Depending of the kernel launch dmension, the async memcpy for
the next batch may be blocked until the previous kernel launch finishes.
This was observed for large dimensions (30 blocks os 512 threads) on a
GTX1050.

Note 2: execution times got slightly worse (but the hope is that it will
be worthy when simulations whose iterations get over a threshold will be
aborted and re-executed at the CPU).

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>