Make max_tau_scatt dynamic again
Current code handles scattering but is not passing the tests with a
references.json extracted from original code with scattering. Maybe one
of the reasons for it is the change, long ago, that fixed max_tau_scatt
for performance issues. As correctness is more important than
performance, this commit reverts the previous change. This makes the
test error smaller (but it still exists) and adds an overhead to the
execution time.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>